+++ src/Tree/TaskTreeBuilderFactory.php
@@ -8,7 +8,7 @@ use Psr\Container\ContainerInterface;
 
 final readonly class TaskTreeBuilderFactory
 {
-    public function __construct(private ContainerInterface $container)
+    public function __construct(private ?ContainerInterface $container)
     {
     }